refactor common components
