+++ src/components/Footer/Footer.js
@@ -3,7 +3,7 @@ import { View } from 'react-native';
 import { connect } from 'react-redux';
 
 import * as actionCreators from '../../actions';
-import { Button } from '../common/Button';
+import { Button } from '../common';
 
 import styles from './styles';
 import { BUTTON_DELETE_TEXT } from './constants';
