+++ src/components/TodoItem/TodoItem.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 
 import * as actionCreators from '../../actions';
 
-import { Button } from '../common/Button';
+import { Button } from '../common';
 import styles from './styles';
 
 class TodoItem extends Component {
