+++ src/components/common/Button/Button.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Text, TouchableOpacity } from 'react-native';
+
+import styles from './styles';
+
+const Button = ({ onPress, style, children, color = '#007aff' }) => {
+  const { buttonStyle, textStyle } = styles;
+
+  return (
+    <TouchableOpacity onPress={onPress} style={[buttonStyle, style, { borderColor: color }]}>
+      <Text style={[textStyle, { color }]}>{children}</Text>
+    </TouchableOpacity>
+  );
+};
+
+export { Button };
