+++ src/components/common/Spinner/Spinner.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, ActivityIndicator } from 'react-native';
 
-import styles from './styles/spinner';
+import styles from './styles';
 
 const Spinner = ({ size }) => (
   <View style={styles.spinnerStyle}>
