+++ src/components/common/Spinner/styles.js
